+++ framework/src/main/groovy/org/moqui/impl/context/NotificationMessageImpl.groovy
@@ -50,6 +50,7 @@ class NotificationMessageImpl implements NotificationMessage, Externalizable {
 
     private NotificationType type = (NotificationType) null
     private Boolean showAlert = (Boolean) null
+    private String emailTemplateId = (String) null
     private Boolean persistOnSend = (Boolean) null
 
     private transient ExecutionContextFactoryImpl ecfiTransient = (ExecutionContextFactoryImpl) null
@@ -128,10 +129,19 @@ class NotificationMessageImpl implements NotificationMessage, Externalizable {
     @Override String getTopic() { topic }
 
     @Override NotificationMessage message(String messageJson) { this.messageJson = messageJson; messageMap = null; return this }
-    @Override NotificationMessage message(Map message) { this.messageMap = Collections.unmodifiableMap(message); messageJson = null; return this }
+    @Override NotificationMessage message(Map message) {
+        this.messageMap = Collections.unmodifiableMap(message) as Map<String, Object>
+        messageJson = null
+        return this
+    }
     @Override String getMessageJson() {
-        if (messageJson == null && messageMap != null)
-            messageJson = JsonOutput.toJson(messageMap)
+        if (messageJson == null && messageMap != null) {
+            try {
+                messageJson = JsonOutput.toJson(messageMap)
+            } catch (Exception e) {
+                logger.warn("Error writing JSON for Notification ${topic} message: ${e.toString()}\n${messageMap}")
+            }
+        }
         return messageJson
     }
     @Override Map<String, Object> getMessageMap() {
@@ -143,18 +153,16 @@ class NotificationMessageImpl implements NotificationMessage, Externalizable {
     @Override NotificationMessage title(String title) { titleTemplate = title; return this }
     @Override String getTitle() {
         if (titleText == null) {
-            if (titleTemplate) {
-                titleText = ecfi.resource.expand(titleTemplate, "", getMessageMap(), true)
-            } else {
-                EntityValue localNotTopic = getNotificationTopic()
-                if (localNotTopic != null) {
-                    if (type == danger && localNotTopic.errorTitleTemplate) {
-                        titleText = ecfi.resource.expand((String) localNotTopic.errorTitleTemplate, "", getMessageMap(), true)
-                    } else if (localNotTopic.titleTemplate) {
-                        titleText = ecfi.resource.expand((String) localNotTopic.titleTemplate, "", getMessageMap(), true)
-                    }
+            EntityValue localNotTopic = getNotificationTopic()
+            if (localNotTopic != null) {
+                if (type == danger && localNotTopic.errorTitleTemplate) {
+                    titleText = ecfi.resource.expand((String) localNotTopic.errorTitleTemplate, "", getMessageMap(), true)
+                } else if (localNotTopic.titleTemplate) {
+                    titleText = ecfi.resource.expand((String) localNotTopic.titleTemplate, "", getMessageMap(), true)
                 }
             }
+            if ((titleText == null || titleText.isEmpty()) && titleTemplate != null && !titleTemplate.isEmpty())
+                titleText = ecfi.resource.expand(titleTemplate, "", getMessageMap(), true)
         }
         return titleText
     }
@@ -202,6 +210,24 @@ class NotificationMessageImpl implements NotificationMessage, Externalizable {
         }
     }
 
+    @Override NotificationMessage emailTemplateId(String id) {
+        emailTemplateId = id
+        if (emailTemplateId != null && emailTemplateId.isEmpty()) emailTemplateId = null
+        return this
+    }
+    @Override String getEmailTemplateId() {
+        if (emailTemplateId != null) {
+            return emailTemplateId
+        } else {
+            EntityValue localNotTopic = getNotificationTopic()
+            if (localNotTopic != null && localNotTopic.emailTemplateId) {
+                return localNotTopic.emailTemplateId
+            } else {
+                return null
+            }
+        }
+    }
+
     @Override NotificationMessage persistOnSend(boolean persist) { persistOnSend = persist; return this }
     @Override boolean isPersistOnSend() {
         if (persistOnSend != null) {
@@ -221,6 +247,7 @@ class NotificationMessageImpl implements NotificationMessage, Externalizable {
         return send()
     }
     @Override NotificationMessage send() {
+        // persist if is persistOnSend
         if (isPersistOnSend()) {
             sentDate = new Timestamp(System.currentTimeMillis())
             // a little trick so that this is available in the closure
@@ -233,7 +260,7 @@ class NotificationMessageImpl implements NotificationMessage, Externalizable {
                                      messageJson:nmi.getMessageJson(), titleText:nmi.getTitle(), linkText:nmi.getLink(),
                                      typeString:nmi.getType(), showAlert:(nmi.showAlert ? 'Y' : 'N')])
                         .disableAuthz().call()
-                // if it's null we got an error
+                // if it's null we got an error so return from closure
                 if (createResult == null) return
 
                 nmi.setNotificationMessageId((String) createResult.notificationMessageId)
@@ -246,6 +273,35 @@ class NotificationMessageImpl implements NotificationMessage, Externalizable {
 
         // now send it to the topic
         ecfi.sendNotificationMessageToTopic(this)
+
+        // send emails if emailTemplateId
+        String localEmailTemplateId = getEmailTemplateId()
+        if (localEmailTemplateId != null && !localEmailTemplateId.isEmpty()) {
+            Set<String> curNotifyUserIds = getNotifyUserIds()
+
+            EntityList emailNotificationUsers = ecfi.entityFacade.find("moqui.security.user.NotificationTopicUser")
+                    .condition("topic", topic).condition("emailNotifications", "Y").disableAuthz().list()
+            int emailNotificationUsersSize = emailNotificationUsers.size()
+            if (emailNotificationUsersSize > 0) {
+                Map<String, Object> wrappedMessageMap = getWrappedMessageMap()
+                
+                for (int i = 0; i < emailNotificationUsersSize; i++) {
+                    EntityValue notificationUser = (EntityValue) emailNotificationUsers.get(i)
+                    String userId = (String) notificationUser.userId
+                    if (!curNotifyUserIds.contains(userId)) continue
+
+                    EntityValue userAccount = ecfi.entityFacade.find("moqui.security.UserAccount")
+                            .condition("userId", userId).disableAuthz().one()
+                    String emailAddress = userAccount?.emailAddress
+                    if (emailAddress) {
+                        // FUTURE: if there is an option to create EmailMessage record also configure emailTypeEnumId (maybe if emailTypeEnumId is set create EmailMessage)
+                        ecfi.serviceFacade.async().name("org.moqui.impl.EmailServices.send#EmailTemplate")
+                                .parameters([emailTemplateId:localEmailTemplateId, toAddresses:emailAddress,
+                                    bodyParameters:wrappedMessageMap, toUserId:userId, createEmailMessage:false]).call()
+                    }
+                }
+            }
+        }
 
         return this
     }
@@ -296,9 +352,20 @@ class NotificationMessageImpl implements NotificationMessage, Externalizable {
         }
     }
 
-    @Override Map<String, Object> getWrappedMessageMap() { [topic:topic, sentDate:sentDate, notificationMessageId:notificationMessageId,
-            message:getMessageMap(), title:getTitle(), link:getLink(), type:getType(), showAlert:isShowAlert()] }
-    @Override String getWrappedMessageJson() { JsonOutput.toJson(getWrappedMessageMap()) }
+    @Override Map<String, Object> getWrappedMessageMap() {
+        EntityValue localNotTopic = getNotificationTopic()
+        return [topic:topic, sentDate:sentDate, notificationMessageId:notificationMessageId, topicDescription:localNotTopic?.description,
+            message:getMessageMap(), title:getTitle(), link:getLink(), type:getType(), showAlert:isShowAlert()]
+    }
+    @Override String getWrappedMessageJson() {
+        Map<String, Object> wrappedMap = getWrappedMessageMap()
+        try {
+            return JsonOutput.toJson(wrappedMap)
+        } catch (Exception e) {
+            logger.warn("Error writing JSON for Notification ${topic} message: ${e.toString()}\n${wrappedMap}")
+            return null
+        }
+    }
 
     void populateFromValue(EntityValue nmbu) {
         this.notificationMessageId = nmbu.notificationMessageId