+++ framework/src/main/groovy/org/moqui/impl/screen/ScreenForm.groovy
@@ -1024,7 +1024,7 @@ class ScreenForm {
                                 addFieldOption(options, fieldNode, childNode, [entry:listOption], ec)
                             } else {
                                 String loString = ObjectUtilities.toPlainString(listOption)
-                                if (loString != null) options.put(loString, loString)
+                                if (loString != null) options.put(loString, ec.l10n.localize(loString))
                             }
                         }
                     }
@@ -1032,7 +1032,7 @@ class ScreenForm {
             } else if ("option".equals(childNode.name)) {
                 String key = ec.resource.expandNoL10n(childNode.attribute('key'), null)
                 String text = ec.resource.expand(childNode.attribute('text'), null)
-                options.put(key, text ?: key)
+                options.put(key, text ?: ec.l10n.localize(key))
             }
         }
         return options
@@ -1064,13 +1064,13 @@ class ScreenForm {
             if (text == null || text.length() == 0) {
                 if (listOptionEvb == null || listOptionEvb.getEntityDefinition().isField("description")) {
                     Object desc = listOption.get("description")
-                    options.put(key, desc != null ? (String) desc : key)
+                    options.put(key, desc != null ? (String) desc : ec.l10n.localize(key))
                 } else {
-                    options.put(key, key)
+                    options.put(key, ec.l10n.localize(key))
                 }
             } else {
                 String value = ec.resource.expand(text, null)
-                if ("null".equals(value)) value = key
+                if ("null".equals(value)) value = ec.l10n.localize(key)
                 options.put(key, value)
             }
         } finally {
@@ -1267,12 +1267,12 @@ class ScreenForm {
             if (validateService) {
                 ServiceDefinition sd = ecfi.serviceFacade.getServiceDefinition(validateService)
                 if (sd == null) throw new BaseArtifactException("Invalid validate-service name [${validateService}] in field [${fieldName}] of form [${location}]")
-                MNode parameterNode = sd.getInParameter((String) fieldNode.attribute('validate-parameter') ?: fieldName)
+                MNode parameterNode = sd.getInParameter((String) subFieldNode.attribute('validate-parameter') ?: fieldName)
                 return parameterNode
             } else if (validateEntity) {
                 EntityDefinition ed = ecfi.entityFacade.getEntityDefinition(validateEntity)
                 if (ed == null) throw new BaseArtifactException("Invalid validate-entity name [${validateEntity}] in field [${fieldName}] of form [${location}]")
-                MNode efNode = ed.getFieldNode((String) fieldNode.attribute('validate-field') ?: fieldName)
+                MNode efNode = ed.getFieldNode((String) subFieldNode.attribute('validate-field') ?: fieldName)
                 return efNode
             }
             return null