+++ framework/src/main/groovy/org/moqui/impl/tools/H2ServerToolFactory.groovy
@@ -55,7 +55,7 @@ class H2ServerToolFactory implements ToolFactory<Server> {
             }
             if (argsString) {
                 String[] args = argsString.split(" ")
-                for (int i = 0; i < args.length; i++) if (args[i].contains('${')) args[i] = SystemBinding.expand(args[i])
+                for (int i = 0; i < args.length; i++) while (args[i].contains('${')) args[i] = SystemBinding.expand(args[i])
                 try {
                     h2Server = Server.createTcpServer(args).start();
                     logger.info("Started H2 remote server on port ${h2Server.getPort()} status: ${h2Server.getStatus()}")